refactoring, add optional db/coll args to get_mongo_client
