+++ pandahub/lib/__init__.py
@@ -1,0 +1,1 @@
+from pandahub.api.internal.mongo_client import mongo_client, get_mongo_client
