+++ pandahub/lib/database_toolbox.py
@@ -1,18 +1,8 @@
 # -*- coding: utf-8 -*-
-from contextlib import contextmanager
-from collections.abc import Iterator
 
 import numpy as np
 import pandas as pd
-from pymongo import MongoClient
-from pymongo.database import Database
-from pymongo.collection import Collection
 
-from pandahub.api.internal.settings import MONGODB_URL, MONGODB_USER, MONGODB_PASSWORD
-from pandahub.api.internal.global_db_client import (
-    __global_mongo_client,
-    _get_mongo_client,
-)
 from pandahub.lib.datatypes import DATATYPES
 import base64
 import hashlib
@@ -431,48 +421,6 @@ def migrate_userdb_to_beanie(ph):
                  {'$unset': 'id'},
                  {'$out': 'users'}]
     userdb.aggregate(migration)
-
-def get_mongo_client(connection_url: str = MONGODB_URL, connection_user: str = MONGODB_USER,
-                     connection_password: str = MONGODB_PASSWORD) -> MongoClient:
-    if __global_mongo_client is None:
-        return _get_mongo_client(connection_url, connection_user, connection_password)
-    else:
-        return __global_mongo_client
-
-@contextmanager
-def mongo_client(database: str | None = None, collection: str | None = None, connection_url: str = MONGODB_URL,
-                 connection_user: str = MONGODB_USER, connection_password: str = MONGODB_PASSWORD) -> Iterator[MongoClient | Database | Collection]:
-    """Contextmanager for pymongo MongoClient / Database / Collection with close after use.
-
-    Parameters
-    ----------
-    database: str or None
-        The database to connect to
-    collection: str or None
-        The collection to connect to
-    connection_url: str or None
-        Defaults to MONGODB_URL env var
-    connection_user: str or None
-        Defaults to MONGODB_USER env var
-    connection_password: str or None
-        Defaults to MONGODB_PASSWORD env var
-
-    Returns
-    -------
-        Contextmanager yielding MongoClient / Database / Collection
-    """
-    if collection is not None and database is None:
-        raise ValueError("Must specify database to access a collection!")
-    client = _get_mongo_client(connection_url, connection_user, connection_password)
-    try:
-        if database is not None and collection is not None:
-            yield client[database][collection]
-        elif database is not None:
-            yield client[database]
-        else:
-            yield client
-    finally:
-        client.close()
 
 
 def get_metadata_for_timeseries_collections(db, data_type=None, net_id=None, element_type=None, element_index=None, **kwargs):
